refactor: new `utilities` and `commands` modules
